+++ shepherd/comm/__init__.py
@@ -1,8 +1,6 @@
 """Package with sockets/greenlets communication helpers."""
 from .messages import *
-from .errors import *
 from .messenger import Messenger
 from .notifier import JobDoneNotifier
 
-__all__ = ['Message', 'InputMessage', 'DoneMessage', 'ErrorMessage', 'MessageError', 'UnexpectedMessageTypeError',
-           'Messenger',  'JobDoneNotifier']
+__all__ = ['Message', 'InputMessage', 'DoneMessage', 'ErrorMessage', 'Messenger',  'JobDoneNotifier']
