+++ shepherd/docker/__init__.py
@@ -1,5 +1,5 @@
-from .errors import DockerError
 from .image import DockerImage
 from .container import DockerContainer
+from .registry import list_images_in_registry
 
-__all__ = ['DockerError', 'DockerContainer', 'DockerImage']
+__all__ = ['DockerContainer', 'DockerImage', 'list_images_in_registry']
