+++ shepherd/docker/registry.py
@@ -1,7 +1,7 @@
 import requests
 from requests.auth import HTTPBasicAuth
 
-from shepherd.shepherd.config import RegistryConfig
+from ..config import RegistryConfig
 
 
 def list_images_in_registry(config: RegistryConfig):
