+++ shepherd/errors/comm.py
