+++ shepherd/sheep/__init__.py
@@ -1,6 +1,6 @@
 from .base_sheep import BaseSheep
 from .docker_sheep import DockerSheep
 from .bare_sheep import BareSheep
-from .errors import SheepError, SheepConfigurationError
+from ..errors.sheep import SheepError, SheepConfigurationError
 
 __all__ = ['BaseSheep',  'DockerSheep', 'BareSheep', 'SheepError', 'SheepConfigurationError']
