+++ shepherd/sheep/docker_sheep.py
@@ -4,9 +4,10 @@ from typing import Dict, Any, Optional, List
 from schematics.types import BooleanType
 
 from .base_sheep import BaseSheep
-from .errors import SheepConfigurationError
-from ..docker import DockerContainer, DockerImage, DockerError
-from ..shepherd.config import RegistryConfig
+from shepherd.errors.sheep import SheepConfigurationError
+from ..docker import DockerContainer, DockerImage
+from ..config import RegistryConfig
+from ..errors.docker import DockerError
 
 
 def extract_gpu_number(device_name: str) -> Optional[str]:
