+++ tests/docker/test_utils.py
@@ -3,7 +3,7 @@ import gevent
 
 import pytest
 
-from shepherd.docker import DockerError
+from shepherd.errors.docker import DockerError
 from shepherd.docker.utils import run_docker_command, kill_blocking_container
 
 from .docker_not_available import docker_not_available
