+++ tests/sheep/conftest.py
@@ -2,8 +2,8 @@ import pytest
 import zmq.green as zmq
 from zmq.error import ZMQError
 
-from shepherd.sheep import BareSheep, DockerSheep, SheepError, SheepConfigurationError
-from shepherd.shepherd.config import RegistryConfig
+from shepherd.sheep import BareSheep, DockerSheep
+from shepherd.config import RegistryConfig
 
 
 @pytest.fixture()
